+++ Mindscape.Raygun4Net.Core/Messages/RaygunMessageDetails.cs
@@ -21,9 +21,9 @@ namespace Mindscape.Raygun4Net.Messages
 
     public RaygunClientMessage Client { get; set; }
 
-    public IList<string> Tags { get; set; }
+    public IList<string> Tags { get; set; } = new List<string>();
 
-    public IDictionary UserCustomData { get; set; }
+    public IDictionary UserCustomData { get; set; } = new Dictionary<string, string>();
 
     public RaygunIdentifierMessage User { get; set; }