+++ Mindscape.Raygun4Net.NetCore.Common/Breadcrumbs/IRaygunBreadcrumbStorage.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+
+namespace Mindscape.Raygun4Net.Breadcrumbs
+{
+  public interface IRaygunBreadcrumbStorage
+  {
+    void Store(RaygunBreadcrumb breadcrumb);
+
+    void Clear();
+
+    int Size();
+    
+    IList<RaygunBreadcrumb> ToList();
+  }
+}