+++ Mindscape.Raygun4Net.NetCore.Common/IRaygunMessageBuilder.cs
@@ -27,6 +27,8 @@ namespace Mindscape.Raygun4Net
     IRaygunMessageBuilder SetRequestDetails(RaygunRequestMessage message);
 
     IRaygunMessageBuilder SetResponseDetails(RaygunResponseMessage message);
+    
+    IRaygunMessageBuilder SetBreadcrumbs(IList<RaygunBreadcrumb> breadcrumbs);
 
     IRaygunMessageBuilder Customise(Action<RaygunMessage> customiseMessage);
   }